Using TensorFlow to predict Jet numbers in CERN Proton Collisions (Evaluator-OMID-BAGHCHEH-SARAEI)
The purpose is to create a machine learning model that can predict the number of Jets in CERN's Proton Collision Dataset, which is stored in CERN_Proton_Collision.zip. I use TensorFlow to accomplish this.

MultiJet primary dataset in AOD format from RunB of 2010. This dataset contains runs from 2010 RunB. The list of all validated runs, which must be applied to all analyses, can be found in http://opendata.cern.ch/record/1000. This file contains events from the MultiJet primary dataset from the CMS open data release, and computes the razor variables MR and Rsq, used in supersymmetric particle searches. More details on the razor variables can be found in Phys. Rev. D 90, 112001